Move query option validation into partial
